Support data change trigger configuration in node
So that the OPC Publisher can be configured to receive a notification
on Status, Status/Value or Status/Value/Timestamp change.

This change adds the field `"OpcNodes[].DataChangeFilter"` that can take
the value "Status", "StatusValue" or "StatusValueTimestamp". This allows
the configuration of the monitored item data change trigger. It can be
left null, in which case the default defined in the OPC UA specs is
"StatusValue".

Fixes #1074