Fix sessions in Twin are not closed when endpoint is deactivated (#1910)
Looks like a regression, the keep alive is counted towards the session's activity state, and any cleanup of it through the collector timer is not happening while keepalive's are being sent (constantly) because the timeout value (1 minute) is much larger than the keep alive timer duration (5 seconds). Also keep alive operations in progress are counted towards the session being active.

Change to only count operations other than keep alive towards "active" state of the session. This has limited risk due to the fact that a reference count is being held by the twin keeping the session active despite its activity state. Once the reference count is released, we want the session GC to kick in and collect the "inactive" session.