Subject: Added active handling for publish errors and session reconnection
This commit introduces a new feature for actively handling reconnecting a session when publishing errors occur due to connectivity issues. A new configuration key `ActivePublishErrorHandlingKey` and its default value `ActivePublishErrorHandlingDefault` have been added in `OpcUaClientConfig.cs`. The corresponding option `ActivePublishErrorHandling` is implemented in `OpcUaClientOptions.cs` and `OpcUaClient.cs`.

The `NotifySessionConnectionState` method has been added to the `IOpcUaSubscription` interface and `OpcUaSubscription.cs` to notify when the session is disconnected. This method has also been added to `OpcUaMonitoredItem.cs` and `OpcUaMonitoredItem.Heartbeat.cs` to update the status code of the last received value when the session is disconnected or reconnected.

The `ActivePublishErrorHandling` option is used in `OpcUaClient.cs` and `OpcUaClientManager.cs` to decide whether to actively handle publishing errors and to initialize the `OpcUaClient` instance respectively.

Lastly, some code related to handling the `BadTooManyOperations` status code in `OpcUaClient.cs` has been removed.