Send uncertain or bad status code heartbeats when session is in error state or disconnected (#2242)
The `NotifySessionConnectionState` method has been added to the `IOpcUaSubscription` interface and `OpcUaSubscription.cs` to notify when the session is disconnected. The heartbeat value status code is updated to a connectivity error state during disconnection.  It is restored upon reconnect or the next good value is received.

This commit introduces a command line option to disable actively handling reconnecting a session when publishing errors occur due to connectivity issues. 

Updated the `buildParameters` in the `azure-pipelines.yml` file to include the repository name, source branch, and commit message. This change enhances the detail of information provided in the build parameters.